+++ source/renderer/app/components/wallet/WalletReceive.js
@@ -111,8 +111,8 @@ export default class WalletReceive extends Component<Props, State> {
     },
   }, {
     options: {
-      validateOnChange: true,
-      validationDebounceWait: 250,
+      validationDebounceWait: 0, // Disable debounce to avoid error state after clearing
+      showErrorsOnClear: false,
     },
   });
 
@@ -123,13 +123,7 @@ export default class WalletReceive extends Component<Props, State> {
         const { spendingPassword } = form.values();
         const password = walletHasPassword ? spendingPassword : null;
         this.props.onGenerateAddress(password);
-
-        // We need to disable on-change validation before reseting the form in order to
-        // avoid debounced validation being called straight after the form is reset
-        form.state.options.set({ validateOnChange: false });
-        form.reset();
-        form.showErrors(false);
-        form.state.options.set({ validateOnChange: true });
+        form.clear();
       },
       onError: () => {}
     });