+++ source/renderer/app/components/loading/Loading.js
@@ -213,19 +213,10 @@ export default class Loading extends Component<Props, State> {
               </div>
             )}
             {isSyncing && (
-              <div className="syncingWrapper">
-                <div className={styles.syncing}>
-                  <h1 className={styles.headline}>
-                    {intl.formatMessage(messages.syncing)} {syncPercentage.toFixed(2)}%
-                  </h1>
-                </div>
-                {(localTimeDifference > allowedTimeDifference) && (
-                  <SystemTimeErrorOverlay
-                    localTimeDifference={localTimeDifference}
-                    currentLocale={currentLocale}
-                    onProblemSolutionClick={onProblemSolutionClick}
-                  />
-                )}
+              <div className={styles.syncing}>
+                <h1 className={styles.headline}>
+                  {intl.formatMessage(messages.syncing)} {syncPercentage.toFixed(2)}%
+                </h1>
               </div>
             )}
             {!isSyncing && !isConnecting && isLoadingDataForNextScreen && (
@@ -235,6 +226,13 @@ export default class Loading extends Component<Props, State> {
                 </h1>
                 <LoadingSpinner />
               </div>
+            )}
+            {(localTimeDifference > allowedTimeDifference) && (
+              <SystemTimeErrorOverlay
+                localTimeDifference={localTimeDifference}
+                currentLocale={currentLocale}
+                onProblemSolutionClick={onProblemSolutionClick}
+              />
             )}
           </div>
         )}