+++ default.nix
@@ -5,7 +5,7 @@ stdenv.mkDerivation {
   name = "daedalus";
 
   buildInputs = [
-    nix bash binutils coreutils curl
+    nix bash binutils coreutils curl gnutar
     git python27 curl electron nodejs-6_x
     nodePackages.node-gyp nodePackages.node-pre-gyp
     gnumake